SceneVisionAI_streamlit/requirements.txt
tensorflow-cpu>=2.15
numpy
pandas
Pillow
streamlit
